+++ src/app/src/test/kotlin/org/vaccineimpact/orderlyweb/tests/unit_tests/models/PermissionSetTests.kt
@@ -71,4 +71,5 @@ class PermissionSetTests: TeamcityTests()
         val sut = PermissionSet(perms)
         assertThat(sut.toString()).isEqualTo("[*/testName, testPrefix:testId/anotherTestName]")
     }
+
 }
